Add task that automatically skips tests for exceptions
A configuration file determines which tests can be skipped. The task
introduced in this commit reads the config and applies labels if test
exception matches with submitted operator.

JIRA: ISV-4382
