+++ ansible/roles/operator-pipeline/templates/openshift/pipelines/community-hosted-pipeline.yml
@@ -272,9 +272,31 @@ spec:
           workspace: repository
           subPath: src
 
-    - name: get-supported-versions
+    - name: apply-test-waivers
+      taskRef:
+        name: apply-test-waivers
+        kind: Task
       runAfter:
         - read-config
+      params:
+        - name: pipeline_image
+          value: "$(params.pipeline_image)"
+        - name: pull_request_url
+          value: "$(params.git_pr_url)"
+        - name: operator_name
+          value: "$(tasks.detect-changes.results.added_operator)"
+        - name: github_token_secret_name
+          value: "$(params.github_token_secret_name)"
+        - name: github_token_secret_key
+          value: "$(params.github_token_secret_key)"
+      workspaces:
+        - name: base
+          workspace: repository
+          subPath: base
+
+    - name: get-supported-versions
+      runAfter:
+        - apply-test-waivers
       taskRef:
         name: get-supported-versions
       params:
