+++ pyproject.toml
@@ -17,6 +17,7 @@ readme = "README.md"
 license = { text = "GPLv3" }
 
 [project.scripts]
+apply-test-waivers = "operatorcert.entrypoints.apply_test_waivers:main"
 bundle-dockerfile = "operatorcert.entrypoints.bundle_dockerfile:main"
 create-container-image = "operatorcert.entrypoints.create_container_image:main"
 create-github-gist = "operatorcert.entrypoints.create_github_gist:main"
